+++ DuoArtifacts/DuoArtifactDefinition.cs
@@ -25,7 +25,7 @@ public sealed class DuoArtifactDefinition
 		new(typeof(DizzyDrakeArtifact), [Deck.dizzy, Deck.eunice], I18n.DizzyDrakeArtifactName, I18n.DizzyDrakeArtifactTooltip, "DizzyDrake", "action.overheat", "status.shieldAlt"),
 		new(typeof(DizzyIsaacArtifact), [Deck.dizzy, Deck.goat], I18n.DizzyIsaacArtifactName, I18n.DizzyIsaacArtifactTooltip, "DizzyIsaac", "action.spawn", new(() => StatusMeta.GetTooltips(Instance.KokoroApi.OxidationStatus.Status, 1)[0]), "status.corrodeAlt"),
 		new(typeof(DizzyMaxArtifact), [Deck.dizzy, Deck.hacker], I18n.DizzyMaxArtifactName, I18n.DizzyMaxArtifactTooltip, "DizzyMax", "status.shieldAlt", new TTGlossary("status.boost", 1)),
-		new(typeof(DizzyPeriArtifact), [Deck.dizzy, Deck.peri], I18n.DizzyPeriArtifactName, I18n.DizzyPeriArtifactTooltip, "DizzyPeri", "status.shieldAlt", "status.overdriveAlt", "status.perfectShield"),
+		new(typeof(DizzyPeriArtifact), [Deck.dizzy, Deck.peri], I18n.DizzyPeriArtifactName, I18n.DizzyPeriArtifactTooltip, "DizzyPeri", "status.shieldAlt", "status.overdriveAlt"),
 		new(typeof(DizzyRiggsArtifact), [Deck.dizzy, Deck.riggs], I18n.DizzyRiggsArtifactName, I18n.DizzyRiggsArtifactTooltip, "DizzyRiggs", "status.shieldAlt", "status.evade"),
 		new(typeof(DrakeIsaacArtifact), [Deck.eunice, Deck.goat], I18n.DrakeIsaacArtifactName, I18n.DrakeIsaacArtifactTooltip, "DrakeIsaac", I18n.HeatAltGlossary),
 		new(typeof(DrakeMaxArtifact), [Deck.eunice, Deck.hacker], I18n.DrakeMaxArtifactName, I18n.DrakeMaxArtifactTooltip, "DrakeMax"),